split functionality from lib.rs into misc.rs
