+++ src/lib.rs
@@ -1,94 +1,10 @@
 use ndarray::ArrayView2;
 
-pub trait Accumulator {
-    fn consume(&mut self, val: f64, weight: f64, bin_idx: usize);
-    fn merge(&mut self, other: &Self);
-}
-
-pub struct Mean {
-    weight: Vec<f64>,
-    total: Vec<f64>,
-}
-
-impl Mean {
-    /// Create a Mean instance with length `n` registers
-    pub fn new(n: usize) -> Result<Mean, &'static str> {
-        if n == 0 {
-            Err("n can't be zero")
-        } else {
-            Ok(Mean {
-                weight: vec![0.0; n],
-                total: vec![0.0; n],
-            })
-        }
-    }
-
-    pub fn get_value(&self, out: &mut [f64], weights_out: &mut [f64]) {
-        for i in 0..self.weight.len() {
-            // TODO need to think about divide by 0
-            out[i] = self.total[i] / self.weight[i];
-
-            // is this the most efficient way to do this?
-            weights_out[i] = self.weight[i];
-        }
-    }
-}
-
-impl Accumulator for Mean {
-    fn consume(&mut self, val: f64, weight: f64, bin_idx: usize) {
-        self.weight[bin_idx] += weight;
-        self.total[bin_idx] += val * weight;
-    }
-
-    fn merge(&mut self, other: &Mean) {
-        for i in 0..self.weight.len() {
-            self.total[i] += other.total[i];
-            self.weight[i] += other.weight[i];
-        }
-    }
-}
-
-pub struct Histogram {
-    hist_bin_edges: Vec<f64>,
-    n_hist_bins: usize,
-    // histogram bins are on the fast axis, distance bins are on the slow axis
-    weight: Vec<f64>,
-}
-
-impl Histogram {
-    pub fn new(n_distance_bins: usize, hist_bin_edges: &[f64]) -> Result<Histogram, &'static str> {
-        if n_distance_bins == 0 {
-            Err("n_distance_bins can't be zero")
-        } else if !hist_bin_edges.is_sorted() {
-            return Err("hist_bin_edges must be sorted (monotonically increasing)");
-        } else {
-            let n_hist_bins = hist_bin_edges.len() - 1;
-            Ok(Histogram {
-                weight: vec![0.0; n_distance_bins * n_hist_bins],
-                hist_bin_edges: hist_bin_edges.to_vec(),
-                n_hist_bins,
-            })
-        }
-    }
-
-    pub fn get_value(&self, out: &mut [f64]) {
-        out.copy_from_slice(&self.weight);
-    }
-}
-
-impl Accumulator for Histogram {
-    fn consume(&mut self, val: f64, weight: f64, dist_bin_idx: usize) {
-        if let Some(hist_bin_idx) = get_bin_idx(val, &self.hist_bin_edges) {
-            self.weight[dist_bin_idx * self.n_hist_bins + hist_bin_idx] += weight;
-        }
-    }
-
-    fn merge(&mut self, other: &Self) {
-        for i in 0..self.weight.len() {
-            self.weight[i] += other.weight[i];
-        }
-    }
-}
+mod misc;
+// pull in symbols that are only used in this file
+use misc::{get_bin_idx, squared_diff_norm};
+// pull in symbols that visible outside of the package
+pub use misc::{Accumulator, Histogram, Mean, diff_norm, dot_product};
 
 /// Collection of point properties.
 ///
@@ -151,79 +67,6 @@ impl<'a> PointProps<'a> {
             1.0
         }
     }
-}
-
-// TODO use binary search, and have specialized version for regularly spaced bins?
-/// Get the index of the bin that the squared distance falls into.
-/// Returns None if its out of bounds.
-fn get_bin_idx(distance_squared: f64, squared_bin_edges: &[f64]) -> Option<usize> {
-    // index of first element greater than distance_squared
-    // (or squared_bin_edges.len() if none are greater)
-    let mut first_greater = 0;
-    for &edge in squared_bin_edges.iter() {
-        if distance_squared < edge {
-            break;
-        }
-        first_greater += 1;
-    }
-    if (first_greater == squared_bin_edges.len()) || (first_greater == 0) {
-        None
-    } else {
-        Some(first_greater - 1)
-    }
-}
-
-/// calculate the squared norm of the difference between two (mathematical) vectors
-/// which are part of rust vecs that encodes a list of vectors with dimension on
-/// the "slow axis"
-fn squared_diff_norm(
-    v1: ArrayView2<f64>,
-    v2: ArrayView2<f64>,
-    i1: usize,
-    i2: usize,
-    n_spatial_dims: usize,
-) -> f64 {
-    let mut sum = 0.0;
-    for k in 0..n_spatial_dims {
-        sum += (v1[[k, i1]] - v2[[k, i2]]).powi(2);
-    }
-    sum
-}
-
-/// computes the euclidean distance between 2 (mathematical) vectors taken from
-/// `values_a` and `values_b`.
-///
-/// # Assumptions
-/// This function assumes that spatial dimension varies along axis 0 of
-/// `values_a` and `values_b` (and that it is the same value for both arrays)
-pub fn diff_norm(
-    values_a: ArrayView2<f64>,
-    values_b: ArrayView2<f64>,
-    i_a: usize,
-    i_b: usize,
-) -> f64 {
-    // TODO come up with a better name for this function
-
-    squared_diff_norm(values_a, values_b, i_a, i_b, values_a.shape()[0]).sqrt()
-}
-
-/// computes a dot product between a (mathematical) vectors taken from
-/// `values_a` and `values_b`.
-///
-/// # Assumptions
-/// This function assumes that spatial dimension varies along axis 0 of
-/// `values_a` and `values_b` (and that it is the same value for both arrays)
-pub fn dot_product(
-    values_a: ArrayView2<f64>,
-    values_b: ArrayView2<f64>,
-    i_a: usize,
-    i_b: usize,
-) -> f64 {
-    let mut sum = 0.0;
-    for k in 0..values_a.shape()[0] {
-        sum += values_a[[k, i_a]] * values_b[[k, i_b]];
-    }
-    sum
 }
 
 fn apply_accum_helper<const CROSS: bool>(
@@ -303,74 +146,4 @@ pub fn apply_accum(
         apply_accum_helper::<true>(accum, points_a, points_a, &squared_bin_edges, pairwise_fn)
     }
     Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    // eventually, we will break up the functionality between files and we
-    // will distribute tests appropriately
-
-    // Accumulator Tests
-    // -----------------
-
-    #[test]
-    fn mean_0len() {
-        assert!(Mean::new(0).is_err());
-    }
-
-    #[test]
-    fn mean_consume_once() {
-        let mut accum = Mean::new(1).unwrap();
-        accum.consume(4.0, 1.0, 0_usize);
-
-        let mut mean_vec = vec![0.0];
-        let mut weight_vec = vec![0.0];
-        accum.get_value(&mut mean_vec, &mut weight_vec);
-
-        assert_eq!(mean_vec[0], 4.0);
-        assert_eq!(weight_vec[0], 1.0);
-    }
-
-    #[test]
-    fn mean_consume_twice() {
-        let mut accum = Mean::new(1).unwrap();
-        accum.consume(4.0, 1.0, 0);
-        accum.consume(8.0, 1.0, 0);
-
-        let mut mean_vec = vec![0.0];
-        let mut weight_vec = vec![0.0];
-        accum.get_value(&mut mean_vec, &mut weight_vec);
-        assert_eq!(mean_vec[0], 6.0);
-        assert_eq!(weight_vec[0], 2.0);
-    }
-
-    #[test]
-    fn merge() {
-        let mut accum = Mean::new(1).unwrap();
-        accum.consume(4.0, 1.0, 0);
-        accum.consume(8.0, 1.0, 0);
-
-        let mut accum_other = Mean::new(1).unwrap();
-        accum_other.consume(1.0, 1.0, 0);
-        accum_other.consume(3.0, 1.0, 0);
-        accum.merge(&accum_other);
-
-        let mut mean_vec = vec![0.0];
-        let mut weight_vec = vec![0.0];
-        accum.get_value(&mut mean_vec, &mut weight_vec);
-        assert_eq!(mean_vec[0], 4.0);
-        assert_eq!(weight_vec[0], 4.0);
-    }
-
-    #[test]
-    fn hist_0len() {
-        assert!(Histogram::new(0, &[0.0, 1.0]).is_err());
-    }
-
-    #[test]
-    fn hist_nonmonotonic() {
-        assert!(Histogram::new(3, &[1.0, 0.0]).is_err());
-    }
 }
